made position a double
also ran the code through the CleanUp.... tool of eclipse

+++ src/main/java/movement/CollisionEngine.java
@@ -10,26 +10,25 @@ import movement.mathDS.Vector.MalformedVectorException;
 
 import java.util.ArrayDeque;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Collections;
 import java.util.HashMap;
 
 
 public class CollisionEngine {
-	
+
 	private static final double teleportOutBuffer = 2;
-	
+
 	private List<Obstacle> obstacleList;
 	private List<Movable> movableList;
-	private Graph<Collidable, float[][]> pointsOfContact;	//directed graph of all the points of contact. Edge values are size 2 arrays of collision locations. First location is collisionlocation of source, second is collisionlocation of dest
-	
+	private Graph<Collidable, double[][]> pointsOfContact;	//directed graph of all the points of contact. Edge values are size 2 arrays of collision locations. First location is collisionlocation of source, second is collisionlocation of dest
+
 	public CollisionEngine(List<Entity> objectList) {
 		setObjectList(objectList);
 	}
 	public CollisionEngine() {
 		this(new ArrayList<Entity>());
 	}
-	
+
 	public void setObjectList(List<Entity> objectList) {
 		var movables = new ArrayList<Movable>();
 		var obstacles = new ArrayList<Obstacle>();
@@ -54,7 +53,7 @@ public class CollisionEngine {
 		obstacleList.remove(object);
 		movableList.remove(object);
 	}
-	
+
 	private List<Movable> getMovableList() {
 		return movableList;
 	}
@@ -73,15 +72,15 @@ public class CollisionEngine {
 		totalList.addAll(getObstacleList());
 		return totalList;
 	}
-	private void setPoC(Graph<Collidable, float[][]> PoC) {
+	private void setPoC(Graph<Collidable, double[][]> PoC) {
 		pointsOfContact = PoC;
 	}
-	private Graph<Collidable, float[][]> getPoC(){
+	private Graph<Collidable, double[][]> getPoC(){
 		return pointsOfContact;
 	}
-	
+
 	public void checkCollisions() throws MalformedVectorException {
-		setPoC(new Graph<Collidable, float[][]>());
+		setPoC(new Graph<Collidable, double[][]>());
 		getPoC().addVertices(getTotalList());
 		for (var object1 : getMovableList()) {
 			collisionDetection(object1);
@@ -100,12 +99,12 @@ public class CollisionEngine {
 			}
 		}
 	}
-	
+
 	private boolean checkBoundsCollision(Movable object1, Collidable object2) {
 		double edge1;
 		double edge2;
-		float[] pos1 = object1.getPosition();
-		float[] pos2 = object2.getPosition();
+		double[] pos1 = object1.getPosition();
+		double[] pos2 = object2.getPosition();
 		double[] dims1 = object1.getDimensions();
 		double[] dims2 = object2.getDimensions();
 		boolean output = true;
@@ -119,38 +118,38 @@ public class CollisionEngine {
 		return output;
 	}
 	private void addToCollisionGraph(Movable object1, Collidable object2) throws MalformedVectorException {
-		var collisionLocation2in1 = object1.getOutline().exactCollisionPosition(object2.getOutline(), relativeLocation(new float[Vector.DIMENSIONS],object1, object2));
-		var collisionLocation1in2 = object2.getOutline().exactCollisionPosition(object1.getOutline(), relativeLocation(new float[Vector.DIMENSIONS],object2, object1));
+		var collisionLocation2in1 = object1.getOutline().exactCollisionPosition(object2.getOutline(), relativeLocation(new double[Vector.DIMENSIONS],object1, object2));
+		var collisionLocation1in2 = object2.getOutline().exactCollisionPosition(object1.getOutline(), relativeLocation(new double[Vector.DIMENSIONS],object2, object1));
 		if (collisionLocation1in2 != null && collisionLocation2in1 != null) {
-			float[][] edge1to2 = {collisionLocation2in1, collisionLocation1in2}; 
-			float[][] edge2to1 = {collisionLocation1in2, collisionLocation2in1}; 
+			double[][] edge1to2 = {collisionLocation2in1, collisionLocation1in2};
+			double[][] edge2to1 = {collisionLocation1in2, collisionLocation2in1};
 			getPoC().addEdge(object1, object2, edge1to2);
 			getPoC().addEdge(object2, object1, edge2to1);
 		}
 	}
-	
-	private void moveableCollision(Movable object1, Movable object2, float[][] collisionLocations) throws MalformedVectorException {
+
+	private void moveableCollision(Movable object1, Movable object2, double[][] collisionLocations) throws MalformedVectorException {
 		var outputForce1 = new Vector();
 		var outputForce2 = new Vector();
-		
+
 		var normal1 = object1.getOutline().getNormal(collisionLocations[0]);
 		var normal2 = object2.getOutline().getNormal(collisionLocations[1]);
-		
-		
+
+
 		var CoR_Effect = (1+(object1.getCoR() * object2.getCoR()))/2; //you gotta average it with 1, because otherwise a value of 0 would violate conservation of momentum. It's actually just how the math works out 2.
 		var timeScaleInverse = 1/Movable.TIMESCALE;
-		
+
 		double massRatio = object1.getMass()/object2.getMass();
-		
+
 		var collisionForce1 = new Vector(CoR_Effect *timeScaleInverse * Vector.getComponentParallel(object1.getVelocity(),normal2)* object1.getMass() * (1+((1-massRatio)/(1+massRatio))),
-										Vector.vectorMovingWith(object1.getVelocity() , normal2) ? Vector.directionOfReverse(normal2) : normal2.getDirection());
+				Vector.vectorMovingWith(object1.getVelocity() , normal2) ? Vector.directionOfReverse(normal2) : normal2.getDirection());
 		var collisionForce2 = new Vector(CoR_Effect * timeScaleInverse* Vector.getComponentParallel(object2.getVelocity(),normal1)* object2.getMass() * (1-((1-massRatio)/(1+massRatio))),
-										Vector.vectorMovingWith(object2.getVelocity() , normal1) ? Vector.directionOfReverse(normal1) : normal1.getDirection());
+				Vector.vectorMovingWith(object2.getVelocity() , normal1) ? Vector.directionOfReverse(normal1) : normal1.getDirection());
 
 		outputForce1.addVector(collisionForce1);
 		outputForce1.addVector(Vector.getReverse(collisionForce2));
 		outputForce2.setComponents(Vector.getReverse(outputForce1).getComponents());
-		
+
 		//assures that if the object gets in, it gets teleported immediately out
 		var relativeEdgeLocation1 = relativeLocation(object1.getOutline().pointOnEdge(collisionLocations[0]), object1, object2);
 		double distance;
@@ -162,21 +161,21 @@ public class CollisionEngine {
 		}
 		for (Obstacle obs : getObstacleList()) {
 			if (checkBoundsCollision(object1, obs)) {
-				var exactObstacleCollisionLocation = object1.getOutline().exactCollisionPosition(obs.getOutline(), relativeLocation(new float[Vector.DIMENSIONS], object1, obs));
+				var exactObstacleCollisionLocation = object1.getOutline().exactCollisionPosition(obs.getOutline(), relativeLocation(new double[Vector.DIMENSIONS], object1, obs));
 				if (exactObstacleCollisionLocation != null) {
 					double distanceInObs = obs.getOutline().distanceIn(relativeLocation(object1.getOutline().pointOnEdge(exactObstacleCollisionLocation), object1, obs));
 					if (distanceInObs > teleportOutBuffer) {
 						var revertVector = new Vector(distanceInObs - teleportOutBuffer, Vector.directionOfReverse(normal2));
 						object1.move(revertVector);
 						object2.move(revertVector);
-					
+
 					}
 				}
 			}
 		}
 		for (Obstacle obs : getObstacleList()) {
 			if (checkBoundsCollision(object2, obs)) {
-				var exactObstacleCollisionLocation = object2.getOutline().exactCollisionPosition(obs.getOutline(), relativeLocation(new float[Vector.DIMENSIONS], object2, obs));
+				var exactObstacleCollisionLocation = object2.getOutline().exactCollisionPosition(obs.getOutline(), relativeLocation(new double[Vector.DIMENSIONS], object2, obs));
 				if (exactObstacleCollisionLocation != null) {
 					double distanceInObs = obs.getOutline().distanceIn(relativeLocation(object2.getOutline().pointOnEdge(exactObstacleCollisionLocation), object2, obs));
 					if (distanceInObs > teleportOutBuffer) {
@@ -184,33 +183,33 @@ public class CollisionEngine {
 						object2.move(revertVector);
 						object1.move(revertVector);
 					}
-				}		
+				}
 			}
 		}
 		object1.applyForce(outputForce1);
 		object2.applyForce(outputForce2);
 	}
-	private void obstacleCollision(Obstacle o, Movable m, float[][] collisionLocations) throws MalformedVectorException {
+	private void obstacleCollision(Obstacle o, Movable m, double[][] collisionLocations) throws MalformedVectorException {
 
 		var normalObstacle = o.getOutline().getNormal(collisionLocations[0]);
-		var outputForce = Vector.vectorMovingWith(m.getVelocity(), normalObstacle) ? 
-							new Vector() :
-							new Vector((1+m.getCoR() * o.getCoR()) *Vector.getComponentParallel(m.getVelocity(), normalObstacle) * (1/Movable.TIMESCALE) *m.getMass()
-									  ,normalObstacle.getDirection());
-		//assures that if the object gets in, it gets teleported immediately out
-		var edgeCollisionLocation = m.getOutline().pointOnEdge(collisionLocations[1]);
-		float[] relativeEdgeCollisionLocation = relativeLocation(edgeCollisionLocation, m, o);double distanceEdgeIn = o.getOutline().distanceIn(relativeEdgeCollisionLocation);
-		if (distanceEdgeIn>teleportOutBuffer ) {
-			m.move(new Vector(distanceEdgeIn-teleportOutBuffer, normalObstacle.getDirection()));
-		}			
-		m.applyForce(outputForce);
+		var outputForce = Vector.vectorMovingWith(m.getVelocity(), normalObstacle) ?
+				new Vector() :
+					new Vector((1+m.getCoR() * o.getCoR()) *Vector.getComponentParallel(m.getVelocity(), normalObstacle) * (1/Movable.TIMESCALE) *m.getMass()
+							,normalObstacle.getDirection());
+				//assures that if the object gets in, it gets teleported immediately out
+				var edgeCollisionLocation = m.getOutline().pointOnEdge(collisionLocations[1]);
+				double[] relativeEdgeCollisionLocation = relativeLocation(edgeCollisionLocation, m, o);double distanceEdgeIn = o.getOutline().distanceIn(relativeEdgeCollisionLocation);
+				if (distanceEdgeIn>teleportOutBuffer ) {
+					m.move(new Vector(distanceEdgeIn-teleportOutBuffer, normalObstacle.getDirection()));
+				}
+				m.applyForce(outputForce);
 	}
 
 	private void organiseCollisionOrder() {	//sorts the movable list, such that they are in the appropriate order to be processed
-		//little bit of terminology. By 'abstraction', I mean the number of objects between an object and a wall. 
+		//little bit of terminology. By 'abstraction', I mean the number of objects between an object and a wall.
 		Map<Movable,int[]> abstractionMap = new HashMap<Movable, int[]>();	//an object with 0 abstraction, relative a wall, either hasn't been measured yet, or isn't connected to it.
 		for (Movable mov : getMovableList()) {
-			abstractionMap.put(mov, new int[getObstacleList().size()]);		
+			abstractionMap.put(mov, new int[getObstacleList().size()]);
 		}
 		for (Obstacle obs : getObstacleList()) {
 			OCOObstacleMethod(obs, abstractionMap, getObstacleList().indexOf(obs));
@@ -226,10 +225,10 @@ public class CollisionEngine {
 		getMovableList().sort((mov1,mov2) -> maxAbstractionMap.get(mov1).compareTo(maxAbstractionMap.get(mov1)));
 		Collections.reverse(getMovableList());
 	}
-	
+
 	private void OCOObstacleMethod(Obstacle obs, Map<Movable, int[]> abstractionMap, int indexOfRootObstacle) {	//this actually has some slight troubles with certain setups that are isomorphic in terms of graph
 		Queue<Movable> movableQueue = new ArrayDeque<Movable>();	//connections, but different in terms of which objects are "on top" of the other. This is an edge case, and will be treated somewhate reasonably.
-		for(Collidable col:getPoC().getVertexConnections(obs)) {	//there is no way to distinguish these cases using a method like this. The method would be required to somehow be aware of the directions of the 
+		for(Collidable col:getPoC().getVertexConnections(obs)) {	//there is no way to distinguish these cases using a method like this. The method would be required to somehow be aware of the directions of the
 			abstractionMap.get(col)[indexOfRootObstacle] = 1;		//forces acting on things and the direction to the wall in question. This would be i)more complex (read: slower), ii) not /super/ useful,
 			movableQueue.add((Movable) col);						// iii) bowing down to specificity
 		}
@@ -237,7 +236,7 @@ public class CollisionEngine {
 			OCOMovableMethod(movableQueue.remove(), abstractionMap, indexOfRootObstacle, movableQueue);
 		}
 	}
-	
+
 	private void OCOMovableMethod(Movable mov, Map<Movable,int[]> abstractionMap, int indexOfRootObstacle, Queue<Movable> movableQueue) {
 		for(Collidable col:getPoC().getVertexConnections(mov)) {
 			if (getMovableList().contains(col) && !movableQueue.contains(col) && abstractionMap.get(col)[indexOfRootObstacle] == 0) {
@@ -246,7 +245,7 @@ public class CollisionEngine {
 			}
 		}
 	}
-	
+
 	private void applyNormalForces() throws MalformedVectorException {	//should only be run after movable list is organised. Requires that the object that is furthest from a wall has it's velocity collision done first
 		var collisionOrder = getMovableList();							//this is because the result of one normal force being applied has an effect on the results of others. Turns out the order of calculation for doing
 		Vector collideeNormal;											//this in one pass is to do the objects with the highest "wall abstraction" first.
@@ -256,8 +255,8 @@ public class CollisionEngine {
 				collideeNormal = collidee.getOutline().getNormal(getPoC().getEdgeValue(collidee, collider)[0]);
 				if (getObstacleList().contains(collidee)){
 					if (Vector.vectorMovingWith(collider.getActiveForce(), collideeNormal)){
-						collider.applyForce(new Vector(Vector.getComponentParallel(collider.getActiveForce(), collideeNormal), collideeNormal.getDirection()));						
-					}	
+						collider.applyForce(new Vector(Vector.getComponentParallel(collider.getActiveForce(), collideeNormal), collideeNormal.getDirection()));
+					}
 				}else if (collisionOrder.indexOf(collider)< collisionOrder.indexOf(collidee)) {
 					if (Vector.vectorMovingWith(collider.getActiveForce(), collideeNormal)){
 						collider.applyForce( new Vector(Vector.getComponentParallel(collider.getActiveForce(), collideeNormal), collideeNormal.getDirection()));
@@ -270,11 +269,11 @@ public class CollisionEngine {
 					}
 				}
 			}
-		}	
+		}
 	}
 
-	private void applyVelocityCollisions() throws MalformedVectorException {	//doesn't care about order. This is because velocity values don't change immediately after a velocity collision, rather, a force is a applied 
-		var collisionOrder = getMovableList();									// which alters the velocity later. In the snapshot of time which this set of collisions takes place in, velocity is constant. 
+	private void applyVelocityCollisions() throws MalformedVectorException {	//doesn't care about order. This is because velocity values don't change immediately after a velocity collision, rather, a force is a applied
+		var collisionOrder = getMovableList();									// which alters the velocity later. In the snapshot of time which this set of collisions takes place in, velocity is constant.
 		for (Movable collider: collisionOrder) {
 			for (Collidable collidee : getPoC().getVerticesConnectedTo(collider)) {
 				if (getObstacleList().contains(collidee)) {
@@ -282,13 +281,13 @@ public class CollisionEngine {
 				} else if (collisionOrder.indexOf(collider) < collisionOrder.indexOf(collidee)){
 					moveableCollision(collider, (Movable)collidee, getPoC().getEdgeValue(collider, collidee));
 				}
-			}		
+			}
 		}
 	}
-	
-	
-	static private float[] relativeLocation(float[] locationRelative1, Entity object1, Entity object2) {
-		float[] locationRelative2 = new float[Vector.DIMENSIONS];
+
+
+	static private double[] relativeLocation(double[] locationRelative1, Entity object1, Entity object2) {
+		double[] locationRelative2 = new double[Vector.DIMENSIONS];
 		for (int i = 0; i<Vector.DIMENSIONS; i++) {
 			locationRelative2[i] = locationRelative1[i] + object1.getPosition()[i] - object2.getPosition()[i];
 		}
